# #144 fix along with some crude parser optimizations