- bytes.Buffer used instead of collecting result in string or []byte
- faster skipping whitespace in parseHexString, relaying str -> hex conversion to base library
- added benchmarks to test suite

Results:
`parseName` 26850ns/op 359alloc => 17464ns/op 95alloc
`parseHexString` 17251ns/op 13alloc => 10700ns/op 10alloc

Other touched functions gain around 5-10%